// frontend/src/app/graphic-builder/shared/enums/visualization-query-params.enum.ts
export enum VisualizationQueryParamsEnum {
  PATTERN_NAME = 'pattern',
  PATTERN_ID = 'patternId',
  VISUALIZATION_ID = 'visualizationId',
  CHART = 'chart',
  /**
   * Define navigation on save visualization
   */
  CALLBACK = 'callback',
  /**
   * Define if run as edit or not
   */
  MODE = 'mode'
}